Create setting editor extension.
