+++ jupyterlab/package.json
@@ -32,6 +32,7 @@
     "@jupyterlab/running-extension": "^0.6.0",
     "@jupyterlab/services": "^0.45.0",
     "@jupyterlab/services-extension": "^0.6.0",
+    "@jupyterlab/settingeditor-extension": "^0.1.0",
     "@jupyterlab/shortcuts-extension": "^0.6.0",
     "@jupyterlab/tabmanager-extension": "^0.6.0",
     "@jupyterlab/terminal-extension": "^0.6.0",
@@ -75,6 +76,7 @@
       "@jupyterlab/rendermime-extension",
       "@jupyterlab/running-extension",
       "@jupyterlab/services-extension",
+      "@jupyterlab/settingeditor-extension",
       "@jupyterlab/shortcuts-extension",
       "@jupyterlab/tabmanager-extension",
       "@jupyterlab/terminal-extension",
