+++ packages/all-packages/package.json
@@ -58,6 +58,7 @@
     "@jupyterlab/running-extension": "^0.6.0",
     "@jupyterlab/services": "^0.45.0",
     "@jupyterlab/services-extension": "^0.6.0",
+    "@jupyterlab/settingeditor-extension": "^0.1.0",
     "@jupyterlab/shortcuts-extension": "^0.6.0",
     "@jupyterlab/tabmanager-extension": "^0.6.0",
     "@jupyterlab/terminal": "^0.6.0",
