+++ packages/settingeditor-extension/package.json
@@ -1,12 +1,13 @@
 {
-  "name": "@jupyterlab/apputils-extension",
-  "version": "0.6.0",
-  "description": "JupyterLab - Application Utilities Extension",
+  "name": "@jupyterlab/settingeditor-extension",
+  "version": "0.1.0",
+  "description": "JupyterLab - Setting Editor Extension",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "files": [
     "lib/*.d.ts",
-    "lib/*.js"
+    "lib/*.js",
+    "style/*.css"
   ],
   "directories": {
     "lib": "lib/"
@@ -14,7 +15,12 @@
   "dependencies": {
     "@jupyterlab/application": "^0.6.0",
     "@jupyterlab/apputils": "^0.6.0",
-    "@jupyterlab/coreutils":  "^0.6.0"
+    "@jupyterlab/codeeditor": "^0.6.0",
+    "@jupyterlab/coreutils":  "^0.6.0",
+    "@phosphor/messaging": "^1.2.0",
+    "@phosphor/signaling": "^1.2.0",
+    "@phosphor/virtualdom": "^1.1.0",
+    "@phosphor/widgets": "^1.2.0"
   },
   "devDependencies": {
     "rimraf": "^2.5.2",
