+++ packages/settingeditor-extension/src/index.ts
@@ -1,0 +1,80 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+import '../style/settingeditor.css';
+
+import {
+  JupyterLab, JupyterLabPlugin
+} from '@jupyterlab/application';
+
+import {
+  ILayoutRestorer, InstanceTracker
+} from '@jupyterlab/apputils';
+
+import {
+  IEditorServices
+} from '@jupyterlab/codeeditor';
+
+import {
+  ISettingRegistry
+} from '@jupyterlab/coreutils';
+
+import {
+  SettingEditor
+} from './settingeditor';
+
+
+/**
+ * The command IDs used by the setting editor.
+ */
+namespace CommandIDs {
+  export
+  const open = 'setting-editor:open';
+};
+
+
+/**
+ * The default setting editor extension.
+ */
+const plugin: JupyterLabPlugin<void> = {
+  activate: (app: JupyterLab, restorer: ILayoutRestorer, registry: ISettingRegistry, editorServices: IEditorServices) => {
+    const { commands, shell } = app;
+    const namespace = 'setting-editor';
+    const factoryService = editorServices.factoryService;
+    const editorFactory = factoryService.newInlineEditor.bind(factoryService);
+    const tracker = new InstanceTracker<SettingEditor>({ namespace });
+
+    // Handle state restoration.
+    restorer.restore(tracker, {
+      command: CommandIDs.open,
+      args: widget => ({ }),
+      name: widget => namespace
+    });
+
+    commands.addCommand(CommandIDs.open, {
+      execute: () => {
+        if (tracker.currentWidget) {
+          shell.activateById(tracker.currentWidget.id);
+          return;
+        }
+
+        const editor = new SettingEditor({ editorFactory, registry });
+
+        tracker.add(editor);
+        editor.id = namespace;
+        editor.title.label = 'Settings';
+        editor.title.closable = true;
+        shell.addToMainArea(editor);
+        shell.activateById(editor.id);
+      },
+      label: 'Settings'
+    });
+  },
+  id: 'jupyter.extensions.setting-editor',
+  requires: [ILayoutRestorer, ISettingRegistry, IEditorServices],
+  autoStart: true
+};
+
+export default plugin;
